requirements.txt
pandas~=2.2.0
streamlit~=1.30.0
yfinance~=0.2.36
python-dateutil~=2.8.2
requests~=2.31.0
plotly~=5.18.0
scipy~=1.12.0
scikit-learn==1.4.1.post1
st-supabase-connection==1.2.2
numpy~=1.26.4
toml~=0.10.2
supabase~=2.4.0
hydralit-components==1.0.10
hydralit==1.0.14
keras==3.1.1
tensorflow==2.16.1
coverage==7.4.4